COVID_Django/requirements.txt
tensorflow==2.6.0
opencv-python==4.5.3.56
scikit-learn==0.24.2
scikit-image==0.17.2
djangorestframework==3.12.4
django-cors-headers==3.10.0
Django==3.2.7
Pillow==8.3.2
gunicorn==20.0.4